// shared/src/expressions/types.d.ts
export type ExpressionValueType = number | boolean | string | Vec2 | ExpressionValueType[];
// ...
export interface BaseNode {
    type: ASTNodeType;
    start: number;
    end: number;
}
export interface NumberLiteralNode extends BaseNode {
    type: 'NumberLiteral';
    value: number;
}
export interface BooleanLiteralNode extends BaseNode {
    type: 'BooleanLiteral';
    value: boolean;
}
export interface StringLiteralNode extends BaseNode {
    type: 'StringLiteral';
    value: string;
}
export interface VectorLiteralNode extends BaseNode {
    type: 'VectorLiteral';
    x: ASTNode;
    y: ASTNode;
}
export interface IdentifierNode extends BaseNode {
    type: 'Identifier';
    name: string;
}
export interface MemberAccessNode extends BaseNode {
    type: 'MemberAccess';
    object: ASTNode;
    property: string;
}
export type BinaryOperator = '+' | '-' | '*' | '/' | '%' | '<' | '<=' | '>' | '>=' | '==' | '!=' | '&&' | '||';
export interface BinaryOpNode extends BaseNode {
    type: 'BinaryOp';
    operator: BinaryOperator;
    left: ASTNode;
    right: ASTNode;
}
export type UnaryOperator = '-' | '!';
export interface UnaryOpNode extends BaseNode {
    type: 'UnaryOp';
    operator: UnaryOperator;
    operand: ASTNode;
}
export interface TernaryNode extends BaseNode {
    type: 'Ternary';
    condition: ASTNode;
    consequent: ASTNode;
    alternate: ASTNode;
}
export interface FunctionCallNode extends BaseNode {
    type: 'FunctionCall';
    name: string;
    args: ASTNode[];
}
export interface TemplateStringNode extends BaseNode {
    type: 'TemplateString';
    parts: Array<{
        type: 'text';
        value: string;
    } | {
        type: 'expr';
        node: ASTNode;
    }>;
}
export type ASTNode = NumberLiteralNode | BooleanLiteralNode | StringLiteralNode | VectorLiteralNode | IdentifierNode | MemberAccessNode | BinaryOpNode | UnaryOpNode | TernaryNode | FunctionCallNode | TemplateStringNode;
export interface EntityContext {
    id: string;
    transform: {
        x: number;
        y: number;
        angle: number;
    };
    velocity?: Vec2;
    health?: number;
    maxHealth?: number;
    [key: string]: unknown;
}
export interface EntityManagerLike {
    getEntity(id: string): {
        id: string;
        transform: {
            x: number;
            y: number;
            scaleX: number;
            scaleY: number;
        };
        visual?: {
            type: string;
            width?: number;
            height?: number;
            radius?: number;
        };
    } | undefined;
    getEntitiesByTag(tag: string): {
        id: string;
        transform: {
            x: number;
            y: number;
            scaleX: number;
            scaleY: number;
        };
        visual?: {
            type: string;
            width?: number;
            height?: number;
            radius?: number;
        };
    }[];
}
export type CustomFunction = (args: ExpressionValueType[], ctx: EvalContext) => ExpressionValueType;
export interface EvalContext {
    time: number;
    wave: number;
    frameId: number;
    dt: number;
    self?: EntityContext;
    variables: Record<string, ExpressionValueType>;
    random: () => number;
    entityManager?: EntityManagerLike;
    customFunctions?: Record<string, CustomFunction>;
}
export interface CompiledExpression<T extends ExpressionValueType = ExpressionValueType> {
    ast: ASTNode;
    source: string;
    evaluate: (ctx: EvalContext) => T;
    dependencies: string[];
}